bella: first step towards contextual conversation
So, bella can now understand context in a conversation, say, play a random
song, then say "this is a good song", bella now knows that the song it just
played is a good song. The important thing to be noted is it can now
understand words such as "this", "it" etc and know what user meant based on
context.