bella: switch to Apache 2.0
The Apache license is quite like the BSD license in its scope. It's pretty free and easy,
you're not required to distribute the source code of a covered work, and you don't have
to worry about it "infecting" a derived work; you just need to be sure to include their license.